+++ lib/eskema.dart
@@ -1,4 +1,5 @@
 library eskema;
 
 export 'validators.dart';
+export 'validator.dart';
 export 'result.dart';
